extract postgres container config to superclass.
