+++ universal-application-tool-0.0.1/test/FunctionalTest.java
@@ -12,10 +12,8 @@ import java.util.Optional;
 import java.util.concurrent.CompletionStage;
 import models.Person;
 import org.junit.Test;
-import play.Application;
 import play.mvc.Http;
 import play.mvc.Result;
-import play.test.WithApplication;
 import play.twirl.api.Content;
 import repository.PersonRepository;
 
@@ -24,24 +22,7 @@ import repository.PersonRepository;
  *
  * <p>https://www.playframework.com/documentation/latest/JavaFunctionalTest
  */
-public class FunctionalTest extends WithApplication {
-
-  protected Application provideApplication() {
-    return fakeApplication(
-        ImmutableMap.of(
-            "db.default.driver",
-            "org.testcontainers.jdbc.ContainerDatabaseDriver",
-            "db.default.url",
-            /* This is a magic string.  The components of it are
-             * jdbc: the standard java database connection uri scheme
-             * tc: Testcontainers - the tool that starts a new container per test.
-             * postgresql: which container to start
-             * 9.6.8: which version of postgres to start
-             * ///: hostless URI scheme - anything here would be ignored
-             * databasename: the name of the db to connect to - any string is okay.
-             */
-            "jdbc:tc:postgresql:9.6.8:///databasename"));
-  }
+public class FunctionalTest extends WithPostgresContainer {
 
   @Test
   public void renderTemplate() {
